requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.4
tzdata==2025.3
whitenoise==6.11.0
